Several fixes, added positional tracking
I’ve made a copy of the example scene, which I can merge back in later
when this is all clean. My intention for my own projects is to use the
GL3 programmable renderer pipeline, so will work on making that usable
too.

There were a few confounding factors in getting this working:

* The scene scale in the included example was huge, at non-physical
size (500 m wide). This meant that even when I did get positional
tracking was working, I didn’t notice it, since any head movements
didn’t give a sense of parallax with distant objects. It also had the
side effect of making any stereo effects un-noticeable since the eye
separation offsets coming from the had are also in physical scale.

* Part of the culprit here is the ofEasyCam which by default translates
the camera quite a large distance away from the origin (based on
viewport size).  This is now disabled with cam.setAutoDistance(false);

* The previous example also had drawing errors due to lack of depth
testing in the Fbo. This is now enabled.